requirements.txt
numpy>=2.4.0
pandas>=2.3.3
mathplotlib>=3.10.8
PyYAML>=6.0.3
